move registerSerivceWorker to src/utils
